Move txnReducer to v2 lib
