+++ v2/components/Mint/package.json
@@ -18,13 +18,13 @@
     "@snx-v2/parseTxnError": "workspace:*",
     "@snx-v2/stakingCalculations": "workspace:*",
     "@snx-v2/txnLink": "workspace:*",
+    "@snx-v2/txnReducer": "workspace:*",
     "@snx-v2/useDebtData": "workspace:*",
     "@snx-v2/useDelegateWallet": "workspace:*",
     "@snx-v2/useExchangeRatesData": "workspace:*",
     "@snx-v2/useFeePoolData": "workspace:*",
     "@snx-v2/useMintMutation": "workspace:*",
     "@snx-v2/useSynthsBalances": "workspace:*",
-    "@snx-v3/txnReducer": "workspace:*",
     "@storybook/react": "^6.5.16",
     "@synthetixio/wei": "^2.74.4",
     "@tanstack/react-query": "^4.24.6",
