+++ v2/components/RewardsItem/package.json
@@ -14,6 +14,7 @@
     "@snx-v2/parseTxnError": "workspace:*",
     "@snx-v2/stakingCalculations": "workspace:*",
     "@snx-v2/txnLink": "workspace:*",
+    "@snx-v2/txnReducer": "workspace:*",
     "@snx-v2/useApr": "workspace:*",
     "@snx-v2/useClaimRewardsMutation": "workspace:*",
     "@snx-v2/useDebtData": "workspace:*",
@@ -24,7 +25,6 @@
     "@snx-v2/useGetLiquidationRewards": "workspace:*",
     "@snx-v2/useRewardsAvailable": "workspace:*",
     "@snx-v2/useSynthsBalances": "workspace:*",
-    "@snx-v3/txnReducer": "workspace:*",
     "@storybook/react": "^6.5.16",
     "@tanstack/react-query": "^4.24.6",
     "date-fns": "^2.29.3",
