+++ v2/lib/txnReducer/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@snx-v3/txnReducer",
+  "name": "@snx-v2/txnReducer",
   "private": true,
   "main": "index.ts",
   "version": "0.0.1"
