+++ v2/lib/useClaimLiquidatorRewardsMutation/package.json
@@ -5,9 +5,9 @@
   "version": "0.0.9",
   "dependencies": {
     "@snx-v2/ContractContext": "workspace:*",
+    "@snx-v2/txnReducer": "workspace:*",
     "@snx-v2/useGasOptions": "workspace:*",
     "@snx-v2/useSynthetixContracts": "workspace:*",
-    "@snx-v3/txnReducer": "workspace:*",
     "@tanstack/react-query": "^4.24.6",
     "react": "^18.2.0"
   }
