+++ v2/lib/useClaimRewardsMutation/package.json
@@ -4,10 +4,10 @@
   "main": "index.ts",
   "version": "0.0.9",
   "dependencies": {
+    "@snx-v2/txnReducer": "workspace:*",
     "@snx-v2/useDelegateWallet": "workspace:*",
     "@snx-v2/useGasOptions": "workspace:*",
     "@snx-v2/useSynthetixContracts": "workspace:*",
-    "@snx-v3/txnReducer": "workspace:*",
     "@tanstack/react-query": "^4.24.6",
     "react": "^18.2.0"
   }
