+++ v2/lib/useClaimRewardsMutation/useClaimRewardsMutation.ts
@@ -2,7 +2,7 @@ import { useReducer } from 'react';
 import { useMutation } from '@tanstack/react-query';
 import { useFeePool } from '@snx-v2/useSynthetixContracts';
 import { useGasOptions } from '@snx-v2/useGasOptions';
-import { initialState, reducer } from '@snx-v3/txnReducer';
+import { initialState, reducer } from '@snx-v2/txnReducer';
 import { DelegateWallet, useDelegateWallet } from '@snx-v2/useDelegateWallet';
 
 const createPopulateTransaction = (
