+++ v2/lib/useSelfLiquidationMutation/package.json
@@ -4,9 +4,9 @@
   "main": "index.ts",
   "version": "0.0.9",
   "dependencies": {
+    "@snx-v2/txnReducer": "workspace:*",
     "@snx-v2/useGasOptions": "workspace:*",
     "@snx-v2/useSynthetixContracts": "workspace:*",
-    "@snx-v3/txnReducer": "workspace:*",
     "@synthetixio/wei": "^2.74.4",
     "@tanstack/react-query": "^4.24.6",
     "react": "^18.2.0"
